% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{A dummy data frame}
\format{A data frame}
\description{
A dummy data frame
}
\examples{
## Code use to generate data
set.seed(1)
dummy_data = as.data.frame(matrix(runif(100), ncol=4))
}
