% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{miniaa}
\alias{miniaa}
\title{miniaa}
\format{A data frame}
\description{
Providers of health care in the USA are made publicly
available by the US government.
The resulting datasets are large (over 4 GB unzipped) and can be accessed from
http://download.cms.gov/nppes/NPI_Files.html.
In this dataset each row is a registered health care provider.
The columns contain information on these providers, including name, address and
telephone number.
Because there are so many column variables (329) much of the data is redundant.
This data set is a sub set of the real data sets
}
