% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{premier}
\alias{premier}
\title{Premier league 2016/2017}
\format{A data frame}
\description{
Information on teams for the 16/17 season, such as summer spend/ net spend (millions)
, final position, % of english players, games til they were safe from relegation, games til they
scored 40 goals.
}
